split FB graph code into separate class
